DATAREDIS-425 - !! DARFT !! Add support for converting java.util.Map.
We now allow mapping of 'java.util.Map' types. The map key must not be a complex object or 'Object.class'.

  class Person {
    String firstname;
    String lastname;
    Map<String, String> attributes;
    Map<String, Person> relatives;
  }

The above is converted to something like

  firstname=rand
  lastname=al’thor
  attributes.[eye-color]=grey
  attributes.[hair-color]=red
  relatives.[father].firstname=janduin

In case the actual Map value type does not match the maps value type declaration the actual type is added as '_class'.

  relatives.[previous-incarnation]._class=com.example.TaVeren
  relatives.[previous-incarnation].firstname=lews
  relatives.[previous-incarnation].lastname=therin telamon