DATAREDIS-425 - !! DARFT !! Store reference data as semantic references.
Store simple and collection like properties as semantic references. 
Please note that referenced objects need to already be present in Redis in order to be resolved correctly. 

  @KeySpace("persons")
  class Person {

    @Id String id;
    String firstname;
    @Reference City location;
  }

  @KeySpace("locations")
  class City {
    @Id String id;
    String name;
  }


  City tarValon = template.insert(new City("1", "Tar Valon"));
  repo.save(new Person("moiraine", tarValon));

is represented as 

  _class=org.example.Person
  id=99bc2ee3-d0bf-4c44-8ef9-b1f486eb554a
  firstname=moiraine
  currentLocation=locations:1

The currentLocation property now no longer holds all properties of City (like 'location.id=1, location.name="Tar Valon"), but consists of a single entry 'currentLocation=locations:1' that points to the element with key 'locations:1' for the 'currentLocation' property. 

Lazy initialization of referenced Objects is at this time not possible but will be part of the story later on.